Clean up unused files and functions
