+++ lib/BotvacRobot.js
@@ -134,11 +134,17 @@ class BotvacRobot {
    async stopAndDock() {
     const self = this;
     const robot = self.robot;
+    
+    //We might need to stop current cleaning cycle before docking
+    try {
+      await this.stopCleaningCycle();
+    } catch (error) {
+      //Do nothing on error. 
+      //This likely means the robot was already paused/stopped and we can try to dock it
+    }
+    
     const stopAndDockPromise = new Promise((resolve, reject) => {
       try {
-        //TODO: we should only pause if the robot is active somehow
-
-        await this.stopCleaningCycle();
         // It can take significant time from BotVac Pause until it can be sent to dock.
         // Resolving this with recursive loop for now that will give it ~30 seconds to resolve
         (async function loop(parentScope) {
